refactor ddb client to pull out of gamedata
