+++ ddb/errors.py
@@ -1,0 +1,8 @@
+class DDBException(Exception):
+    """Base exception for all DDB-related exceptions"""
+    pass
+
+
+class AuthException(DDBException):
+    """Something happened during auth that shouldn't have"""
+    pass
